Added superuser checks back in

+++ mySite/Test/views_passage.py
@@ -9,8 +9,8 @@ from .views_detail import *
 
 
 def addPassageEditor(request):
-    #if not request.user.is_superuser: # TODO: add back in
-    #    return HttpResponse("You are not a superuser")
+    if not request.user.is_superuser:
+        return HttpResponse("You are not a superuser")
     submit_url = reverse('Test:addPassage')
     form_data = {}
     header = "Add a passage"
@@ -18,8 +18,8 @@ def addPassageEditor(request):
     html = editor(request,submit_url,form_data,header,initial_text) # See MCEditor.views
     return html
 def addPassage(request):
-    #if not request.user.is_superuser:
-    #    return HttpResponse("You are not a superuser")
+    if not request.user.is_superuser:
+        return HttpResponse("You are not a superuser")
     form_text = request.POST.get("form-text")
     passage = Passage(text=form_text)
     passage.save()
@@ -29,8 +29,8 @@ def addPassage(request):
     return recurseQuestionContainerDetail(request, questionContainer)
 
 def editPassageEditor(request,passage_pk):
-    #if not request.user.is_superuser:
-    #    return HttpResponse("You are not a superuser")
+    if not request.user.is_superuser:
+        return HttpResponse("You are not a superuser")
     submit_url = reverse('Test:editPassage',args=[passage_pk])
     form_data = {}
     header = "Edit a passage"
@@ -41,8 +41,8 @@ def editPassageEditor(request,passage_pk):
     html = editor(request,submit_url,form_data,header,initial_text) # See MCEditor.views
     return html
 def editPassage(request,passage_pk):
-    #if not request.user.is_superuser:
-    #    return HttpResponse("You are not a superuser")
+    if not request.user.is_superuser:
+        return HttpResponse("You are not a superuser")
     form_text = request.POST.get("form-text")
     passage = Passage.objects.get(pk=passage_pk)
     passage.text = form_text
@@ -50,8 +50,8 @@ def editPassage(request,passage_pk):
     return recurseQuestionContainerDetail(request, passage)
 
 def deletePassage(request,passage_pk):
-    #if not request.user.is_superuser:
-    #    return HttpResponse("You are not a superuser")
+    if not request.user.is_superuser:
+        return HttpResponse("You are not a superuser")
     passage = Passage.objects.get(pk=passage_pk)
     passage.delete()
     return HttpResponseRedirect( reverse('Test:questionContainerList') )
